build: harden CI and CMake for v1.0.0
- Bump version to 1.0.0
- Add Ninja generator to all CI pipelines
- Default COBRA_BUILD_TESTS and COBRA_BUILD_LLVM_PASS to OFF
- Gate GoogleTest behind COBRA_BUILD_TESTS in superbuild
- Fix C++ standard mismatch: superbuild deps now build as C++23
- Fix CMAKE_CXX_FLAGS overwrite with CMAKE_CXX_FLAGS_INIT
- Fix cobra-verify z3 linkage (PRIVATE → PUBLIC) and add PIC
- Add LLVM_FOUND guard in CobraPass CMakeLists
- Add find_dependency for absl/hwy in CobraConfig.cmake.in
- Add -Wall -Wextra -Wpedantic compiler warnings
- Add USE_EXTERNAL_HIGHWAY option for system highway
- Move forwarding configs to cmake/config/ templates
- Split lint into independent CI job
- Add LLVM script SHA256 verification
- Add compiler/tool version logging step
- Hash-based cache key from dependencies/*.cmake
- Parallel ctest with nproc/sysctl fallback
- Update macOS runner to macos-15
- Split dependabot groups (infrastructure vs tools)
- Update BUILD.md prerequisites to C++23
- Update README dependency comments and optional test instructions

+++ lib/core/GhostResidualSolver.cpp
@@ -61,10 +61,9 @@ namespace cobra {
                 5, 6,  11, 14, 19, 22, 29, 34, 41, 50, 59, 66, 73, 82, 91, 100,
                 9, 12, 17, 20, 25, 30, 35, 40, 47, 56, 63, 70, 79, 86, 95, 102,
             };
-            for (int p = 0; p < kNumProbes; ++p) {
+            for (size_t p = 0; p < kNumProbes; ++p) {
                 for (uint32_t v = 0; v < num_vars && v < 6; ++v) {
-                    bank[static_cast< size_t >(p)].values[v] =
-                        kSeeds[static_cast< size_t >((p * 6) + v)] & kMask;
+                    bank[p].values[v] = kSeeds[(p * 6) + v] & kMask;
                 }
             }
         }
